Change file name to reflect move from lambdas to functors.
