+++ singularity-eos/eos/get_sg_eos_functors.hpp
